requirements.txt
celery==5.5.3
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
groq==1.0.0
langchain==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
pinecone==8.0.0
PyMuPDF==1.26.6
python-dotenv==1.2.1
redis==7.0.1